requirements.txt
# 3Proxy + HAProxy Stack — Python Dependencies
PyYAML>=6.0.1
Jinja2>=3.1.3
psutil>=5.9.8
aiohttp>=3.9.3
prometheus-client>=0.20.0
